Split network into submodule
