+++ src/oemof/solph/network/bus.py
@@ -1,0 +1,43 @@
+# -*- coding: utf-8 -*-
+
+"""Classes used to model energy supply systems within solph.
+
+Classes are derived from oemof core network classes and adapted for specific
+optimization tasks. An energy system is modelled as a graph/network of nodes
+with very specific constraints on which types of nodes are allowed to be
+connected.
+
+SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
+SPDX-FileCopyrightText: Simon Hilpert
+SPDX-FileCopyrightText: Cord Kaldemeyer
+SPDX-FileCopyrightText: Stephan Günther
+SPDX-FileCopyrightText: Birgit Schachler
+
+SPDX-License-Identifier: MIT
+
+"""
+
+from oemof.network import network as on
+
+from oemof.solph import blocks
+
+
+class Bus(on.Bus):
+    """A balance object. Every node has to be connected to Bus.
+
+    Notes
+    -----
+    The following sets, variables, constraints and objective parts are created
+     * :py:class:`~oemof.solph.blocks.Bus`
+
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.balanced = kwargs.get("balanced", True)
+
+    def constraint_group(self):
+        if self.balanced:
+            return blocks.Bus
+        else:
+            return None
