+++ src/oemof/solph/network/energy_system.py
@@ -1,0 +1,46 @@
+# -*- coding: utf-8 -*-
+
+"""Classes used to model energy supply systems within solph.
+
+Classes are derived from oemof core network classes and adapted for specific
+optimization tasks. An energy system is modelled as a graph/network of nodes
+with very specific constraints on which types of nodes are allowed to be
+connected.
+
+SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
+SPDX-FileCopyrightText: Simon Hilpert
+SPDX-FileCopyrightText: Cord Kaldemeyer
+SPDX-FileCopyrightText: Stephan Günther
+SPDX-FileCopyrightText: Birgit Schachler
+
+SPDX-License-Identifier: MIT
+
+"""
+
+from oemof.network import energy_system as es
+
+
+class EnergySystem(es.EnergySystem):
+    """A variant of :class:`EnergySystem
+    <oemof.core.energy_system.EnergySystem>` specially tailored to solph.
+
+    In order to work in tandem with solph, instances of this class always use
+    `solph.GROUPINGS <oemof.solph.GROUPINGS>`. If custom groupings are
+    supplied via the `groupings` keyword argument, `solph.GROUPINGS
+    <oemof.solph.GROUPINGS>` is prepended to those.
+
+    If you know what you are doing and want to use solph without
+    `solph.GROUPINGS <oemof.solph.GROUPINGS>`, you can just use
+    :class:`core's EnergySystem <oemof.core.energy_system.EnergySystem>`
+    directly.
+    """
+
+    def __init__(self, **kwargs):
+        # Doing imports at runtime is generally frowned upon, but should work
+        # for now. See the TODO in :func:`constraint_grouping
+        # <oemof.solph.groupings.constraint_grouping>` for more information.
+        from oemof.solph.groupings import GROUPINGS
+
+        kwargs["groupings"] = GROUPINGS + kwargs.get("groupings", [])
+
+        super().__init__(**kwargs)
