+++ src/oemof/solph/network/flow.py
@@ -19,38 +19,10 @@ SPDX-License-Identifier: MIT
 
 from warnings import warn
 
-from oemof.network import energy_system as es
 from oemof.network import network as on
 from oemof.tools import debugging
 
-from oemof.solph import blocks
 from oemof.solph.plumbing import sequence
-
-
-class EnergySystem(es.EnergySystem):
-    """A variant of :class:`EnergySystem
-    <oemof.core.energy_system.EnergySystem>` specially tailored to solph.
-
-    In order to work in tandem with solph, instances of this class always use
-    `solph.GROUPINGS <oemof.solph.GROUPINGS>`. If custom groupings are
-    supplied via the `groupings` keyword argument, `solph.GROUPINGS
-    <oemof.solph.GROUPINGS>` is prepended to those.
-
-    If you know what you are doing and want to use solph without
-    `solph.GROUPINGS <oemof.solph.GROUPINGS>`, you can just use
-    :class:`core's EnergySystem <oemof.core.energy_system.EnergySystem>`
-    directly.
-    """
-
-    def __init__(self, **kwargs):
-        # Doing imports at runtime is generally frowned upon, but should work
-        # for now. See the TODO in :func:`constraint_grouping
-        # <oemof.solph.groupings.constraint_grouping>` for more information.
-        from oemof.solph.groupings import GROUPINGS
-
-        kwargs["groupings"] = GROUPINGS + kwargs.get("groupings", [])
-
-        super().__init__(**kwargs)
 
 
 class Flow(on.Edge):
@@ -242,131 +214,3 @@ class Flow(on.Edge):
                 "Investment flows cannot be combined with "
                 + "nonconvex flows!"
             )
-
-
-class Bus(on.Bus):
-    """A balance object. Every node has to be connected to Bus.
-
-    Notes
-    -----
-    The following sets, variables, constraints and objective parts are created
-     * :py:class:`~oemof.solph.blocks.Bus`
-
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.balanced = kwargs.get("balanced", True)
-
-    def constraint_group(self):
-        if self.balanced:
-            return blocks.Bus
-        else:
-            return None
-
-
-class Sink(on.Sink):
-    """An object with one input flow."""
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        check_node_object_for_missing_attribute(self, "inputs")
-
-    def constraint_group(self):
-        pass
-
-
-class Source(on.Source):
-    """An object with one output flow."""
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        check_node_object_for_missing_attribute(self, "outputs")
-
-    def constraint_group(self):
-        pass
-
-
-class Transformer(on.Transformer):
-    """A linear Transformer object with n inputs and n outputs.
-
-    Parameters
-    ----------
-    conversion_factors : dict
-        Dictionary containing conversion factors for conversion of each flow.
-        Keys are the connected bus objects.
-        The dictionary values can either be a scalar or an iterable with length
-        of time horizon for simulation.
-
-    Examples
-    --------
-    Defining an linear transformer:
-
-    >>> from oemof import solph
-    >>> bgas = solph.Bus(label='natural_gas')
-    >>> bcoal = solph.Bus(label='hard_coal')
-    >>> bel = solph.Bus(label='electricity')
-    >>> bheat = solph.Bus(label='heat')
-
-    >>> trsf = solph.Transformer(
-    ...    label='pp_gas_1',
-    ...    inputs={bgas: solph.Flow(), bcoal: solph.Flow()},
-    ...    outputs={bel: solph.Flow(), bheat: solph.Flow()},
-    ...    conversion_factors={bel: 0.3, bheat: 0.5,
-    ...                        bgas: 0.8, bcoal: 0.2})
-    >>> print(sorted([x[1][5] for x in trsf.conversion_factors.items()]))
-    [0.2, 0.3, 0.5, 0.8]
-
-    >>> type(trsf)
-    <class 'oemof.solph.network.Transformer'>
-
-    >>> sorted([str(i) for i in trsf.inputs])
-    ['hard_coal', 'natural_gas']
-
-    >>> trsf_new = solph.Transformer(
-    ...    label='pp_gas_2',
-    ...    inputs={bgas: solph.Flow()},
-    ...    outputs={bel: solph.Flow(), bheat: solph.Flow()},
-    ...    conversion_factors={bel: 0.3, bheat: 0.5})
-    >>> trsf_new.conversion_factors[bgas][3]
-    1
-
-    Notes
-    -----
-    The following sets, variables, constraints and objective parts are created
-     * :py:class:`~oemof.solph.blocks.Transformer`
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        check_node_object_for_missing_attribute(self, "inputs")
-        check_node_object_for_missing_attribute(self, "outputs")
-
-        self.conversion_factors = {
-            k: sequence(v)
-            for k, v in kwargs.get("conversion_factors", {}).items()
-        }
-
-        missing_conversion_factor_keys = (
-            set(self.outputs) | set(self.inputs)
-        ) - set(self.conversion_factors)
-
-        for cf in missing_conversion_factor_keys:
-            self.conversion_factors[cf] = sequence(1)
-
-    def constraint_group(self):
-        return blocks.Transformer
-
-
-def check_node_object_for_missing_attribute(obj, attribute):
-    if not getattr(obj, attribute):
-        msg = (
-            "Attribute <{0}> is missing in Node <{1}> of {2}.\n"
-            "If this is intended and you know what you are doing you can"
-            "disable the SuspiciousUsageWarning globally."
-        )
-        warn(
-            msg.format(attribute, obj.label, type(obj)),
-            debugging.SuspiciousUsageWarning,
-        )
